Overhaul of DICOM_to_internal_source Feature List:     1) Reworked DICOM input to read in data sets with several scan sets over time, which are then sorted by time and then z height.  The data sets are then all added together for the activity file.
    2) To go with feature (1), if DICOM tag (0054,1001)
reads "BQML", then the activity read from the DICOM file is
scaled by the scan duration and the size of voxel (0,0,0),
so that the activity file represents the total decays per
voxel, and adding several data sets together using feature
(1) will account for the different scan times.

    3) By default, the top and bottom 2 z slices are omitted
from the activity file due to the noisiness of the data and
the relative lack of importance in those regions.  The
number of omitted slices can also be changed using the flag
omitOuterSlices=x as an input, where x is the number of
slices to omit from each side.

    4) The default settings for filterLowDensity and
filterLowActivity are set to 0.5 (g/cm^3) and 0.01%,
respectively.

    5) A new input option, outputName=S, is introduced to
change the name out of the output file to S.txt.